JavaFx TERMINADO, Modificacion del main 40% completado +/-
