+++ nbproject/project.properties
@@ -19,8 +19,12 @@ build.test.results.dir=${build.dir}/test/results
 #debug.transport=dt_socket
 debug.classpath=\
     ${run.classpath}
+debug.modulepath=\
+    ${run.modulepath}
 debug.test.classpath=\
     ${run.test.classpath}
+debug.test.modulepath=\
+    ${run.test.modulepath}
 # Files in build.classes.dir which should be excluded from distribution jar
 dist.archive.excludes=
 # This directory is removed when the project is cleaned:
@@ -38,6 +42,8 @@ javac.classpath=\
 javac.compilerargs=
 javac.deprecation=false
 javac.external.vm=true
+javac.modulepath=
+javac.processormodulepath=
 javac.processorpath=\
     ${javac.classpath}
 javac.source=1.8
@@ -47,11 +53,14 @@ javac.test.classpath=\
     ${build.classes.dir}:\
     ${libs.junit_4.classpath}:\
     ${libs.hamcrest.classpath}
+javac.test.modulepath=\
+    ${javac.modulepath}
 javac.test.processorpath=\
     ${javac.test.classpath}
 javadoc.additionalparam=
 javadoc.author=false
 javadoc.encoding=${source.encoding}
+javadoc.html5=false
 javadoc.noindex=false
 javadoc.nonavbar=false
 javadoc.notree=false
@@ -60,21 +69,24 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-main.class=controller.App
+jlink.launcher=false
+jlink.launcher.name=Elias_jeanpierr_Reto0
+main.class=controller.Main
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
-platform.active=default_platform
+platform.active=JDK_1.8
 run.classpath=\
     ${javac.classpath}:\
     ${build.classes.dir}
-# Space-separated list of JVM arguments used when running the project.
-# You may also define separate properties like run-sys-prop.name=value instead of -Dname=value.
-# To set system properties for unit tests define test-sys-prop.name=value:
-run.jvmargs=
+run.modulepath=\
+    ${javac.modulepath}:\
+    ${libs.JavaFx19.classpath}
 run.test.classpath=\
     ${javac.test.classpath}:\
     ${build.test.classes.dir}
+run.test.modulepath=\
+    ${javac.test.modulepath}
 source.encoding=UTF-8
 src.dir=src
 test.src.dir=test
