+++ src/controller/Main.java
@@ -1,0 +1,26 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/javafx/FXMain.java to edit this template
+ */
+package controller;
+
+import vista.ViewImplementation;
+import vista.ViewImplementationText;
+
+/**
+ *
+ * @author DarkD
+ */
+public class Main {
+    
+    /**
+    /**
+     * @param args the command line arguments
+     */
+    public static void main(String[] args) {
+    
+          new ViewImplementation().showSaludo();
+          new ViewImplementationText().showSaludo();
+    }
+    
+}
