--- src/controller/ViewFXMLController.java
@@ -1,38 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package controller;
-
-import java.net.URL;
-import java.util.ResourceBundle;
-import javafx.fxml.FXML;
-import javafx.fxml.Initializable;
-import javafx.scene.control.Label;
-import model.ModelImplements;
-/**
- * FXML Controller class
- *
- * @author 2dam
- */
-public class ViewFXMLController implements Initializable {
-
-    @FXML
-    private Label LabelMessage;
-    
-    ModelImplements moderlImp = new ModelImplements();
-
-    /**
-     * Initializes the controller class.
-     */
-    @Override
-    public void initialize(URL url, ResourceBundle rb) {
-        // TODO
-    }    
-    
- /*public void setMessage(){
-      LabelMessage.setText(moderlImp.getSaludo());
-       
-   }*/
-}
