+++ src/model/config.properties
@@ -1,6 +1,6 @@
 URL = jdbc:mysql://localhost:3306/reto0?serverTimezone=Europe/Madrid&useSSL=false
 USER = root
-PASS = abcd*1234
+PASS = root
 
 # Tipo de visualizacion
 # 1 = texto
