--- src/vista/ViewFXML.fxml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-
-
-<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="600.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8.0.171" fx:controller="controller.ViewFXMLController">
-   <children>
-      <Label id="LabelMessage" fx:id="LabelMessage" layoutX="229.0" layoutY="165.0" text="Label">
-         <font>
-            <Font name="Arial Black" size="47.0" />
-         </font>
-      </Label>
-   </children>
-</AnchorPane>
