+++ src/vista/Viewfx.fxml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.text.Font?>
+
+<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/18" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.ViewfxController">
+   <children>
+      <Label id="labelMessage" fx:id="labelMessage" layoutX="108.0" layoutY="175.0" prefHeight="50.0" prefWidth="384.0" text="Label" textAlignment="CENTER" AnchorPane.bottomAnchor="175.0" AnchorPane.leftAnchor="108.0" AnchorPane.rightAnchor="108.0" AnchorPane.topAnchor="175.0">
+         <font>
+            <Font name="Arial Black" size="33.0" />
+         </font>
+      </Label>
+   </children>
+</AnchorPane>
